+++ docs/.vitepress/config.ts
@@ -1,10 +1,10 @@
 /*
  * @Author: weisheng
  * @Date: 2023-07-27 10:26:09
- * @LastEditTime: 2023-12-07 15:08:32
+ * @LastEditTime: 2024-01-06 22:20:13
  * @LastEditors: weisheng
  * @Description: 
- * @FilePath: \wot-design-uni\docs\.vitepress\config.ts
+ * @FilePath: /wot-design-uni/docs/.vitepress/config.ts
  * 记得注释
  */
 import { defineConfig } from 'vitepress';
@@ -231,10 +231,13 @@ export default defineConfig({
             text: "DatetimePickerView 时间选择器视图"
           }, {
             link: "/component/form",
-            text: "Form 表单组合"
+            text: "Form 表单"
           }, {
             link: "/component/input",
             text: "Input 输入框"
+          }, {
+            link: "/component/textarea",
+            text: "Textarea 文本域"
           }, {
             link: "/component/input-number",
             text: "InputNumber 计数器"
@@ -265,6 +268,9 @@ export default defineConfig({
           }, {
             link: "/component/upload",
             text: "Upload 上传"
+          }, {
+            link: "/component/password-input",
+            text: "PasswordInput 密码输入框"
           }]
         }, {
           text: "反馈",