+++ docs/.vitepress/theme/components/frame.vue
@@ -4,8 +4,9 @@
 
 <script setup lang="ts">
 import { useRoute, useData } from 'vitepress';
-import { computed, onMounted, ref, watch } from 'vue'
+import { computed, watch } from 'vue'
 const baseUrl = process.env.NODE_ENV === 'production' ? `${location.origin}/demo/?timestamp=${new Date().getTime()}#/` : 'http://localhost:5173/demo/#/'
+
 const route = useRoute()
 const href = computed(() => {
   const path = route.path