+++ src/uni_modules/wot-design-uni/components/wd-col-picker/wd-col-picker.vue
@@ -10,18 +10,23 @@
       >
         <view
           v-if="label || useLabelSlot"
-          :class="`wd-col-picker__label ${required && 'is-required'} ${customLabelClass}`"
+          :class="`wd-col-picker__label ${isRequired && 'is-required'} ${customLabelClass}`"
           :style="labelWidth ? 'min-width:' + labelWidth + ';max-width:' + labelWidth + ';' : ''"
         >
           <block v-if="label">{{ label }}</block>
           <slot v-else name="label"></slot>
         </view>
-        <view
-          :class="`wd-col-picker__value ${ellipsis && 'is-ellipsis'} ${customValueClass} ${showValue ? '' : 'wd-col-picker__value--placeholder'}`"
-        >
-          {{ showValue || placeholder || '请选择' }}
+        <view class="wd-col-picker__body">
+          <view class="wd-col-picker__value-wraper">
+            <view
+              :class="`wd-col-picker__value ${ellipsis && 'is-ellipsis'} ${customValueClass} ${showValue ? '' : 'wd-col-picker__value--placeholder'}`"
+            >
+              {{ showValue || placeholder || '请选择' }}
+            </view>
+            <wd-icon v-if="!disabled && !readonly" custom-class="wd-col-picker__arrow" name="arrow-right" />
+          </view>
+          <view v-if="errorMessage" class="wd-col-picker__error-message">{{ errorMessage }}</view>
         </view>
-        <wd-icon v-if="!disabled && !readonly" custom-class="wd-col-picker__arrow" name="arrow-right" />
       </view>
     </view>
     <wd-action-sheet
@@ -31,6 +36,7 @@
       :close-on-click-modal="closeOnClickModal"
       :z-index="zIndex"
       :safe-area-inset-bottom="safeAreaInsetBottom"
+      @open="handlePickerOpend"
       @close="handlePickerClose"
     >
       <view class="wd-col-picker__selected">
@@ -89,9 +95,11 @@ export default {
 </script>
 
 <script lang="ts" setup>
-import { getCurrentInstance, onMounted, ref, watch } from 'vue'
+import { computed, getCurrentInstance, onMounted, ref, watch } from 'vue'
 import { debounce, getRect, getType } from '../common/util'
 import { useCell } from '../composables/useCell'
+import { FORM_KEY, type FormItemRule } from '../wd-form/types'
+import { useParent } from '../composables/useParent'
 
 const $container = '.wd-col-picker__selected-container'
 const $item = '.wd-col-picker__selected-item'
@@ -132,6 +140,8 @@ interface Props {
   zIndex?: number
   safeAreaInsetBottom?: boolean
   ellipsis?: boolean
+  prop?: string
+  rules?: FormItemRule[]
 }
 
 const props = withDefaults(defineProps<Props>(), {
@@ -157,7 +167,8 @@ const props = withDefaults(defineProps<Props>(), {
   zIndex: 15,
   safeAreaInsetBottom: true,
   ellipsis: false,
-  labelWidth: '33%'
+  labelWidth: '33%',
+  rules: () => []
 })
 
 const pickerShow = ref<boolean>(false)
@@ -207,12 +218,12 @@ watch(
       console.error('[wot design] error(wd-col-picker): the columns props of wd-col-picker should be a two-dimensional array')
       return
     }
-
     if (newValue.length === 0 && !oldValue) return
 
     const newSelectedList = newValue.slice(0)
 
     selectList.value = newSelectedList
+
     selectShowList.value = pickerColSelected.value.map((item, colIndex) => {
       return getSelectedItem(item, colIndex, newSelectedList)[props.labelKey]
     })
@@ -268,6 +279,31 @@ watch(
   }
 )
 
+const { parent: form } = useParent(FORM_KEY)
+
+// 表单校验错误信息
+const errorMessage = computed(() => {
+  if (form && props.prop && form.errorMessages && form.errorMessages[props.prop]) {
+    return form.errorMessages[props.prop]
+  } else {
+    return ''
+  }
+})
+
+// 是否展示必填
+const isRequired = computed(() => {
+  let formRequired = false
+  if (form && form.props.rules) {
+    const rules = form.props.rules
+    for (const key in rules) {
+      if (Object.prototype.hasOwnProperty.call(rules, key) && key === props.prop && Array.isArray(rules[key])) {
+        formRequired = rules[key].some((rule: FormItemRule) => rule.required)
+      }
+    }
+  }
+  return props.required || props.rules.some((rule) => rule.required) || formRequired
+})
+
 const emit = defineEmits(['close', 'update:modelValue', 'confirm'])
 
 onMounted(() => {
@@ -282,6 +318,10 @@ function open() {
 function close() {
   handlePickerClose()
 }
+function handlePickerOpend() {
+  updateLineAndScroll(false)
+}
+
 function handlePickerClose() {
   pickerShow.value = false
   // 如果目前用户正在选择，需要在popup关闭时将数据重置回上次数据，popup 关闭时间 250
@@ -305,9 +345,6 @@ function showPicker() {
   pickerShow.value = true
   lastPickerColSelected.value = pickerColSelected.value.slice(0)
   lastSelectList.value = selectList.value.slice(0)
-  setTimeout(() => {
-    updateLineAndScroll()
-  }, 30)
 }
 
 function getSelectedItem(value, colIndex, selectList) {